# 6005CEM OAS-1.0.0.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Movie Database API",
    "description": "This is a Movie Database API",
    "contact": {
      "email": "akinpelud@uni.coventry.ac.uk"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://virtserver.swaggerhub.com/jordaniel3/6005CEM/1.0.0",
      "description": "SwaggerHub API Auto Mocking"
    },
    {
      "url": "https://tictaclatin-stormkermit-3000.codio-box.uk/api/v1/",
      "description": "6005CEM Movie Database"
    }
  ],
...
  "paths": {
    "/login": {
      "post": {
        "tags": [
          "Visitors"
        ],
        "summary": "Logs users in",
        "description": "Logs User In",
        "operationId": "login",
        "requestBody": {
          "description": "Create new user",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "user created"
          },
          "400": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "Admins"
        ],
        "summary": "Returns all registered users",
        "description": "This path will return all registered users\n",
        "operationId": "allUsers",
        "responses": {
          "200": {
            "description": "All users were returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/allUsers"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Action was not authorised"
          }
        }
      },
      "post": {
        "tags": [
          "Visitors"
        ],
        "summary": "adds a new user to the database",
        "description": "Adds an user to the database",
        "operationId": "addUser",
        "requestBody": {
          "description": "Create new user",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createUser"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "user created"
          },
          "400": {
            "description": "an existing item already exists"
          },
          "403": {
            "description": "invalid input, user not created"
          }
        }
      }
    },
    "/users/:id": {
      "get": {
        "tags": [
          "Admins",
          "Users"
        ],
        "summary": "Returns the corresponding user",
        "description": "This path will return the user belonging to that id (Users can only view their own id path)\n",
        "operationId": "getUserByID",
        "responses": {
          "200": {
            "description": "The user was successfully returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/userByID"
                }
              }
            }
          },
          "304": {
            "description": "Resource has not been modified"
          },
          "403": {
            "description": "Action was not authorised"
          }
        }
      },
      "put": {
        "tags": [
          "Admins",
          "Users"
        ],
        "summary": "updates user info",
        "description": "This path will update the user belonging to that id (Users can only update their own id path)",
        "operationId": "updateUser",
        "requestBody": {
          "description": "Update user",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createUser"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "user updated"
          },
          "400": {
            "description": "Invalid request body"
          },
          "403": {
            "description": "Not authorised"
          }
        }
      },
      "delete": {
        "tags": [
          "Admins",
          "Users"
        ],
        "summary": "Deletes the corresponding user",
        "description": "This path will delete the user belonging to that id (Users can only del their own id path)\n",
        "operationId": "deleteUserByID",
        "responses": {
          "410": {
            "description": "User deleted"
          },
          "403": {
            "description": "Action was not authorised"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "allUsers": {
        "type": "array",
        "example": [
          {
            "id": 1,
            "firstName": "John",
            "lastName": "Smith",
            "username": "Jsmith"
          },
          {
            "id": 2,
            "firstName": "Jon",
            "lastName": "Smith",
            "username": "Jsmith2"
          }
        ],
        "items": {
          "$ref": "#/components/schemas/allUsers_inner"
        }
      },
      "userByID": {
        "required": [
          "firstName",
          "id",
          "lastName",
          "username"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Smith"
          },
          "userName": {
            "type": "string",
            "example": "Jsmith"
          }
        }
      },
      "createUser": {
        "required": [
          "firstName",
          "lastName",
          "password",
          "username"
        ],
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "example": "password"
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Smith"
          },
          "userName": {
            "type": "string",
            "example": "Jsmith"
          }
        }
      },
...
      "allUsers_inner": {
        "required": [
          "firstName",
          "id",
          "lastName",
          "username"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          }
        }
      },
...
    }
  }
}